+++ apps/client_android_modelviewlabsorg_desafio_api/lib/src/utils/consts.dart
@@ -1,4 +1,0 @@
-
-final class Consts {
-  static const String apiKey = String.fromEnvironment('apiKey');
-}