+++ modules/core/lib/src/features/consumo_api_ping/datasource/get_connect_api_datasource.dart/api_ping_provider.dart
@@ -1,0 +1,16 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+import 'package:dependencies/dependencies.dart';
+
+class ApiPingProvider extends GetConnect {
+  final String apiEndpoint;
+  final String apiPing;
+  ApiPingProvider({
+    required this.apiEndpoint,
+    required this.apiPing,
+  });
+
+
+  Future<Response> getPong() => get('$apiEndpoint$apiPing');
+
+
+}