+++ modules/core/lib/src/features/consumo_api_ping/domain/usecase/model/data_api_response_ping.dart
@@ -1,0 +1,9 @@
+final class DataApiResponsePing {
+  final String response;
+  DataApiResponsePing({
+    required this.response,
+  });
+
+  @override
+  String toString() => 'DataApiResponsePing(response: $response)';
+}