+++ modules/core/lib/src/utils/parameters.dart
@@ -1,0 +1,23 @@
+import 'package:dependencies/dependencies.dart';
+
+import 'erros.dart';
+
+final class ParametrosRandom implements ParametersReturnResult {
+  @override
+  final ApiRandomError error;
+
+  ParametrosRandom({
+    required this.error,
+  });
+}
+
+final class ParametrosValidate implements ParametersReturnResult {
+  final String password;
+  @override
+  final ApiValidateError error;
+
+  ParametrosValidate({
+    required this.error,
+    required this.password,
+  });
+}