+++ modules/home/lib/src/ui/valida_senha/valida_senha_page.dart
@@ -28,14 +28,14 @@ class ValidaSenhaPage extends GetView<ValidaSenhaController> {
               height: 80,
               child: IconButton(
                   onPressed: () {
-                    // controller.logOut(
-                    //   onSuccess: () {
-                    //     Get.snackbar("Sucesso", "Sucesso ao fazer LogOut");
-                    //   },
-                    //   onFail: () {
-                    //     Get.snackbar("Erro", "Erro ao fazer LogOut");
-                    //   },
-                    // );
+                    controller.consumoApiPing(
+                      onSuccess: () {
+                        Get.snackbar("Sucesso", "Sucesso ao Chamar API");
+                      },
+                      onFail: () {
+                        Get.snackbar("Erro", "Erro ao Chamar API");
+                      },
+                    );
                   },
                   icon: const FaIcon(FontAwesomeIcons.arrowRightFromBracket)),
             ),
@@ -46,19 +46,15 @@ class ValidaSenhaPage extends GetView<ValidaSenhaController> {
               height: 80,
               child: IconButton(
                   onPressed: () {
-                    // mostrarDialogoConfirmacao(onSuccess: () {
-                    //   controller.apagarConta(
-                    //     confirmacao: true,
-                    //     onSuccess: () {
-                    //       Get.snackbar("Sucesso", "Sucesso ao apagar a conta");
-                    //       Get.toNamed("/login");
-                    //     },
-                    //     onFail: () {
-                    //       Get.snackbar("Erro", "Erro ao apagar a conta");
-                    //     },
-                    //   );
-                    //   Get.back();
-                    // });
+                    controller.consumoApiRandom(
+                      onSuccess: () {
+                        Get.snackbar(
+                            "Sucesso", "Sucesso ao Gerar senha pela API");
+                      },
+                      onFail: () {
+                        Get.snackbar("Erro", "Erro ao Gerar senha pela API");
+                      },
+                    );
                   },
                   icon: const FaIcon(FontAwesomeIcons.userXmark)),
             ),
@@ -67,7 +63,51 @@ class ValidaSenhaPage extends GetView<ValidaSenhaController> {
             child: SizedBox(
               width: 80,
               height: 80,
-              child: Obx(() => Text(controller.teste.value)),
+              child: IconButton(
+                  onPressed: () {
+                    controller.consumoApiValidator(
+                      onSuccess: () {
+                        Get.snackbar(
+                            "Sucesso", "Sucesso ao Validar a senha pela API");
+                      },
+                      onFail: () {
+                        Get.snackbar(
+                            "Erro", "Erro ao Validar a senha pela API");
+                      },
+                      password: 't8p#oOue45',
+                    );
+                  },
+                  icon: const FaIcon(FontAwesomeIcons.angellist)),
+            ),
+          ),
+          Center(
+            child: SizedBox(
+              width: 80,
+              height: 80,
+              child: Obx(() => Text(controller.isPong.value.toString())),
+            ),
+          ),
+          Center(
+            child: SizedBox(
+              width: 80,
+              height: 80,
+              child: Obx(() => Text(controller.radomPassword.value)),
+            ),
+          ),
+          Center(
+            child: SizedBox(
+              width: 200,
+              height: 200,
+              child:
+                  Obx(() => Text(controller.validatePassword.value['message'].toString())),
+            ),
+          ),
+          Center(
+            child: SizedBox(
+              width: 200,
+              height: 200,
+              child:
+                  Obx(() => Text(controller.validatePassword.value['errors'].toString())),
             ),
           ),
         ],