+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.2
+Desenvolvimento do código fonte.
+1 - Conficuração dos and points para consumo da api.
+2 - Criação dos Models.
+3 - Criação da feature e testes para consumo da API endpoint pong.
+4 - Criação da feature e testes para consumo da API endpoint random.
+5 - Criação da feature e testes para consumo da API endpoint validate.
+
 ## 0.0.1
 Configuração inicial do projeto.
 1 - Instalação Dependencies Module.